+++ src/ResourceManager/PolicyInsights/ChangeLog.md
@@ -18,6 +18,9 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Public release of Policy Insights cmdlets
+    - Use API version 2018-04-04
+    - Add PolicyDefinitionReferenceId to the results of Get-AzureRmPolicyStateSummary
 
 ## Version 0.9.1-preview
 * Set minimum dependency of module to PowerShell 5.0